extract Reference object to its own class #9
